+++ src/main/java/de/codeinfection/quickwango/AntiGuest/Preventions/Bukkit/AfkPrevention.java
@@ -4,7 +4,6 @@ import de.codeinfection.quickwango.AntiGuest.AntiGuestBukkit;
 import de.codeinfection.quickwango.AntiGuest.Prevention;
 import java.util.HashMap;
 import org.bukkit.configuration.Configuration;
-import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -43,11 +42,11 @@ public class AfkPrevention extends Prevention
     }
 
     @Override
-    public void enable(final ConfigurationSection config)
+    public void enable()
     {
-        super.enable(config);
+        super.enable();
         this.scheduler = getPlugin().getServer().getScheduler();
-        this.timeout = config.getInt("timeout") * 20;
+        this.timeout = getConfig().getInt("timeout") * 20;
 
         this.trackerMap = new HashMap<Player, PlayerAfkTracker>();
     }